Now prices and refunds are supported for item skin Ids. (If less than 0, the item skin Id will be set based on the entity skin Id)
Support NoEscape plugin, and remove the built-in RaidBlocker. (It was changed in previous versions, but I forgot to write it in the change log)
In target entity detection, when a sub-object is detected, the sub-object is displayed in priority. (e.g. when you look at hemp, you will get the hemp instead of the plant box)
Add options: "Show Message When Enabled/Disabled" and "Display Names (Refund & Price)" (Used to display the name of a special skin)
Remove options: "RemoveType - Normal - Check stash under the foundation" and "Display Names Of Other Things" (It doesn't do much, it just makes the config large)
The "OnRemovableEntityCheckOrPay" and "OnRemovableEntityGiveRefund" hooks add the skinId parameter.